Moving windowing to rendering engine
beginning support for vulkan and d3d12 support may not be implemented

+++ project/RenderEngine/Assets/Text.cpp
@@ -78,7 +78,7 @@ void Text::setText(char * text, size_t length, GLfloat x, GLfloat y, GLfloat sca
 			if (yAdv)
 			{
 				y += f->getFontSize();
-				oldYAdv = f->getFontSize();
+				oldYAdv = (float)f->getFontSize();
 			} else {
 				y += oldYAdv;
 			}
